[FEAT]: SocketProvidable 기능별로 프로토콜 분리
